refactor: name changes from Swagger to OpenAPi
